Further implementation of DOMController
Updated Css to display the new components and the
sidebar. Had some difficulty rotating the text
within the confines of the sidebar.
Corrected fix in the automatic shipplacement so
that we have ships in both orientations
gameboard.placeShip now places ships with the
correct size instead of one to big
DOMController now displays the grid of the first
player. Basic routine for next player should be
about done as well.